\name{read.collapsing.matrix}
\alias{read.collapsing.matrix}
\title{
Read gene/sample varaint count (collapsing) matrix.
}
\description{
This is a helper function that parses the input gene/sample varaint count (collapsing) matrix into an R matrix.
}
\usage{
read.collapsing.matrix(matrix_file)
}
\arguments{
  \item{matrix_file}{The input collapsing matrix.
}
}
\details{
Describe what a collapsing matrix looks like.
}
\value{
Returns a matrix that represents the collapsing matrix, with columns for samples and rows for genes.
}
