\name{read.gene.sets}
\alias{read.gene.sets}
\title{
Read the gene set definitition from input text file
}
\description{
This function read the gene set to be used for mega collapsing analysis.
}
\usage{
read.gene.sets(meta_gene_file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{meta_gene_file}{The input file that defines gene sets.
}
}
\details{
Need to add more details about the gene set format we are using here and the minimal information we will need to create the geneset information.
}
\value{
Return a list of gene sets.
}
