\name{read.seed.list}
\alias{read.seed.list}
\title{
A helper function to read "seed genes"
}
\description{
Forget what this is supposed to be. Will need to investigate if we need it or not, or if we need to re-implement it to fit our purpose.
}
\usage{
read.seed.list(seed_set)
}
\arguments{
  \item{seed_set}{An input for  "seed set" of genes.
}
}
\details{
Need to figure how this is used (or not).
}
\value{
Return a list of "seed set"
}
